[nnx] add pmap
