+++ flax/nnx/__init__.py
@@ -106,18 +106,20 @@ from .nnx.training import optimizer as optimizer
 from .nnx.training.metrics import Metric as Metric
 from .nnx.training.metrics import MultiMetric as MultiMetric
 from .nnx.training.optimizer import Optimizer as Optimizer
-from .nnx.transforms import Jit as Jit
-from .nnx.transforms import Remat as Remat
-from .nnx.transforms import Scan as Scan
-from .nnx.transforms import Vmap as Vmap
-from .nnx.transforms import grad as grad
-from .nnx.transforms import jit as jit
-from .nnx.transforms import remat as remat
-from .nnx.transforms import scan as scan
-from .nnx.transforms import value_and_grad as value_and_grad
-from .nnx.transforms import vmap as vmap
-from .nnx.transforms import eval_shape as eval_shape
-from .nnx.transforms import cond as cond
+from .nnx.transforms.transforms import Jit as Jit
+from .nnx.transforms.transforms import Remat as Remat
+from .nnx.transforms.looping import Scan as Scan
+from .nnx.transforms.parallelization import Vmap as Vmap
+from .nnx.transforms.parallelization import Pmap as Pmap
+from .nnx.transforms.transforms import grad as grad
+from .nnx.transforms.transforms import jit as jit
+from .nnx.transforms.transforms import remat as remat
+from .nnx.transforms.looping import scan as scan
+from .nnx.transforms.transforms import value_and_grad as value_and_grad
+from .nnx.transforms.parallelization import vmap as vmap
+from .nnx.transforms.parallelization import pmap as pmap
+from .nnx.transforms.transforms import eval_shape as eval_shape
+from .nnx.transforms.transforms import cond as cond
 from .nnx.variables import EMPTY as EMPTY
 from .nnx.variables import A as A
 from .nnx.variables import BatchStat as BatchStat
